# fix(query): optimized passwords and secrets query