--- a/assets/queries/common/passwords_and_secrets_in_infrastructure_code/query.rego
+++ b/assets/queries/common/passwords_and_secrets_in_infrastructure_code/query.rego
@@ -4,50 +4,60 @@ import data.generic.common as commonLib
 
 # search for harcoded secrets by looking for their values with a special chars and length
 CxPolicy[result] {
-	docs := input.document[id]
-	treatedDoc = checkName(docs, name, docs[name])
-	keyDoc = treatedDoc[y][father]
+	docs := input.document[_]
 
-	clearParse := ["playbooks", "tasks", "command"]
-	not commonLib.equalsOrInArray(clearParse, father)
+	[path, value] = walk(docs)
+	is_string(value)
+	value = replaceUniCode(value)
+	checkObject := prepare_object(key, value)
 
-	#get all string values from json
-	allValues = regex.find_n("\"[^\"]+\"\\s*:\\s*\"[^\"]+\"[]\n\r,}]", json.marshal(keyDoc), -1)
-
-	correctStrings := getCorrectStrings(replaceUniCode(allValues[m]), father)
-
-	checkforvulnerability(correctStrings)
-
+	check_vulnerability(checkObject)
 	result := {
 		"documentId": docs.id,
-		"searchKey": sprintf("{{%s}}.%s={{%s}}", [correctStrings.id, correctStrings.key, correctStrings.value]),
+		"searchKey": sprintf("%s", [concat(".", path)]),
 		"issueType": "RedundantAttribute",
 		"keyExpectedValue": "Hardcoded secret key should not appear in source",
-		"keyActualValue": correctStrings.value,
+		"keyActualValue": value,
 	}
 }
 
-isUnderPasswordKey(p) {
+prepare_object(key, value) = obj {
+	contains(value, "=")
+
+	obj := {
+		"key": path[minus(count(path), 1)],
+		"value": value,
+	}
+}
+
+prepare_object(key, value) = obj {
+	obj := {
+		"key": path[minus(count(path), 1)],
+		"value": value,
+	}
+}
+
+is_under_password_key(p) {
 	ar = {"pas", "psw", "pwd"}
 	contains(lower(p), ar[_])
 }
 
-isUnderSekretKey(p) = res {
+is_under_secret_key(p) = res {
 	ar = {"secret", "encrypt", "credential"}
 	res := contains(lower(p), ar[_])
 }
 
 #search for default passwords
-checkforvulnerability(correctStrings) {
+check_vulnerability(correctStrings) {
 	commonLib.isDefaultPassword(correctStrings.value)
-	isUnderPasswordKey(correctStrings.key)
+	is_under_password_key(correctStrings.key)
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
 #search for non-default passwords under known names
-checkforvulnerability(correctStrings) {
+check_vulnerability(correctStrings) {
 	#remove short strings
 	count(correctStrings.value) > 4
 	count(correctStrings.value) < 30
@@ -55,14 +65,14 @@ checkforvulnerability(correctStrings) {
 	#password should contain alpha and numeric and not contain spaces
 	count(regex.find_n("[a-zA-Z0-9]+", correctStrings.value, -1)) > 0
 	count(regex.find_n("^[^{{]+$", correctStrings.value, -1)) > 0
-	isUnderPasswordKey(correctStrings.key)
+	is_under_password_key(correctStrings.key)
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
 #search for non-default passwords with upper, lower chars and digits
-checkforvulnerability(correctStrings) { #ignore ascii cases
+check_vulnerability(correctStrings) { #ignore ascii cases
 	#remove short strings
 	count(correctStrings.value) > 6
 	count(correctStrings.value) < 20
@@ -74,20 +84,20 @@ checkforvulnerability(correctStrings) { #ignore ascii cases
 	count(regex.find_n("^[^\\s_]+", correctStrings.value, -1)) > 0
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
 #search for harcoded secrets with known prefixes
-checkforvulnerability(correctStrings) {
+check_vulnerability(correctStrings) {
 	#look for a known prefix
 	contains(correctStrings.value, "PRIVATE KEY")
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
 #search for harcoded secret keys under known names
-checkforvulnerability(correctStrings) {
+check_vulnerability(correctStrings) {
 	#remove short strings
 	count(correctStrings.value) > 8
 
@@ -95,17 +105,17 @@ checkforvulnerability(correctStrings) {
 	count(regex.find_n("^[^\\s$]+$", correctStrings.value, -1)) > 0
 
 	#look for a known names
-	isUnderSekretKey(correctStrings.key)
+	is_under_secret_key(correctStrings.key)
 
 	#remove string with non-keys characters
 	count(regex.find_n("^[^\\s/:@,.-_|]+$", correctStrings.value, -1)) > 0
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
 #search for harcoded secrets by looking for their values with a special chars and length
-checkforvulnerability(correctStrings) {
+check_vulnerability(correctStrings) {
 	#remove short strings
 	count(correctStrings.value) > 30
 
@@ -113,10 +123,10 @@ checkforvulnerability(correctStrings) {
 	count(regex.find_n("^[^\\s/:@,.-_|]+$", correctStrings.value, -1)) > 0
 
 	#remove common key and values
-	checkCommon(correctStrings)
+	check_common(correctStrings)
 }
 
-checkCommon(correctStrings) {
+check_common(correctStrings) {
 	#remove common values
 	not commonLib.isCommonValue(correctStrings.value)
 
@@ -130,6 +140,7 @@ replaceUniCode(allValues) = treatedValue {
 	treatedValue = replace(treatedValue_first, "\\u003e", ">")
 }
 
+#this will be removed, but still needs as example
 #construct correct strings based on dockerfile or other platform
 getCorrectStrings(allValues, name) = correctStrings {
 	#other platforms
@@ -156,16 +167,4 @@ getCorrectStrings(allValues, name) = correctStrings {
 		"value": substring(allStrings.value, 0, count(allStrings.value) - 2),
 		"id": name,
 	}
-}
-
-# clear parser added feilds
-checkName(document, name, docName) = treatedName {
-	clearAns := ["playbooks", "tasks"]
-	commonLib.equalsOrInArray(clearAns, name)
-	treatedName := [x | x := docName[int]]
-} else = treatedName {
-	name == "command"
-	treatedName := [x | x := docName]
-} else = treatedName {
-	treatedName := [x | x := document]
 }