--- a/assets/queries/common/passwords_and_secrets_in_infrastructure_code/test/negative4.tf
+++ b/assets/queries/common/passwords_and_secrets_in_infrastructure_code/test/negative4.tf
@@ -5,9 +5,6 @@ resource "google_container_cluster" "primary" {
   initial_node_count = 3
 
   master_auth {
-    username = ""
-    password = ""
-
     client_certificate_config {
       issue_client_certificate = true
     }