--- a/assets/queries/common/passwords_and_secrets_in_infrastructure_code/test/positive5.dockerfile
+++ b/assets/queries/common/passwords_and_secrets_in_infrastructure_code/test/positive5.dockerfile
@@ -1,3 +1,7 @@
 FROM baseImage
 
 ARG password=pass!1213Fs
+
+
+FROM test2
+ARG password=pass!1213Fs