Signed-off-by: Felipe Avelar <felipe.avelar@checkmarx.com>

Closes #3058 

**Proposed Changes**
- Optimize query

I submit this contribution under the Apache-2.0 license.