@felipe-avelar 2021-05-04
Using `opa bench` to compare, we can see the following improvement on performance:

Running kics scan -p . (kics folder)

--> New query

```
+-------------------------------------------+---------------+
| samples                                   |             1 |
| ns/op                                     |    7730055000 | 7.73s
| B/op                                      |    3986459192 |
| allocs/op                                 |      68875751 |
| histogram_timer_rego_query_eval_ns_75%    |    7730017400 |
| histogram_timer_rego_query_eval_ns_90%    |    7730017400 |
| histogram_timer_rego_query_eval_ns_95%    |    7730017400 |
| histogram_timer_rego_query_eval_ns_99%    |    7730017400 |
| histogram_timer_rego_query_eval_ns_99.9%  |    7730017400 |
| histogram_timer_rego_query_eval_ns_99.99% |    7730017400 |
| histogram_timer_rego_query_eval_ns_count  |          1.00 |
| histogram_timer_rego_query_eval_ns_max    |    7730017400 |
| histogram_timer_rego_query_eval_ns_mean   |    7730017400 |
| histogram_timer_rego_query_eval_ns_median |    7730017400 |
| histogram_timer_rego_query_eval_ns_min    |    7730017400 |
| histogram_timer_rego_query_eval_ns_stddev |             0 |
+-------------------------------------------+---------------+
```
--> Old query
```
+-------------------------------------------+---------------+
| samples                                   |             1 |
| ns/op                                     |   53299056500 | 53.30s
| B/op                                      |   18989655952 |
| allocs/op                                 |     329803520 |
| histogram_timer_rego_query_eval_ns_75%    |   53299000800 |
| histogram_timer_rego_query_eval_ns_90%    |   53299000800 |
| histogram_timer_rego_query_eval_ns_95%    |   53299000800 |
| histogram_timer_rego_query_eval_ns_99%    |   53299000800 |
| histogram_timer_rego_query_eval_ns_99.9%  |   53299000800 |
| histogram_timer_rego_query_eval_ns_99.99% |   53299000800 |
| histogram_timer_rego_query_eval_ns_count  |          1.00 |
| histogram_timer_rego_query_eval_ns_max    |   53299000800 |
| histogram_timer_rego_query_eval_ns_mean   |   53299000800 |
| histogram_timer_rego_query_eval_ns_median |   53299000800 |
| histogram_timer_rego_query_eval_ns_min    |   53299000800 |
| histogram_timer_rego_query_eval_ns_stddev |             0 |
+-------------------------------------------+---------------+
```
-> Larger Repo
-->New query
```
+-------------------------------------------+---------------+
| samples                                   |             1 |
| ns/op                                     |   92831469700 |92.83s
| B/op                                      |   33928882208 |
| allocs/op                                 |     593658279 |
| histogram_timer_rego_query_eval_ns_75%    |   92831384400 |
| histogram_timer_rego_query_eval_ns_90%    |   92831384400 |
| histogram_timer_rego_query_eval_ns_95%    |   92831384400 |
| histogram_timer_rego_query_eval_ns_99%    |   92831384400 |
| histogram_timer_rego_query_eval_ns_99.9%  |   92831384400 |
| histogram_timer_rego_query_eval_ns_99.99% |   92831384400 |
| histogram_timer_rego_query_eval_ns_count  |          1.00 |
| histogram_timer_rego_query_eval_ns_max    |   92831384400 |
| histogram_timer_rego_query_eval_ns_mean   |   92831384400 |
| histogram_timer_rego_query_eval_ns_median |   92831384400 |
| histogram_timer_rego_query_eval_ns_min    |   92831384400 |
| histogram_timer_rego_query_eval_ns_stddev |             0 |
+-------------------------------------------+---------------+
```
--> Old query
```
+-------------------------------------------+---------------+
| samples                                   |             1 |
| ns/op                                     |  777429701200 |777.429s
| B/op                                      |  260060757176 |
| allocs/op                                 |    4518867209 |
| histogram_timer_rego_query_eval_ns_75%    |  777429622100 |
| histogram_timer_rego_query_eval_ns_90%    |  777429622100 |
| histogram_timer_rego_query_eval_ns_95%    |  777429622100 |
| histogram_timer_rego_query_eval_ns_99%    |  777429622100 |
| histogram_timer_rego_query_eval_ns_99.9%  |  777429622100 |
| histogram_timer_rego_query_eval_ns_99.99% |  777429622100 |
| histogram_timer_rego_query_eval_ns_count  |          1.00 |
| histogram_timer_rego_query_eval_ns_max    |  777429622100 |
| histogram_timer_rego_query_eval_ns_mean   |  777429622100 |
| histogram_timer_rego_query_eval_ns_median |  777429622100 |
| histogram_timer_rego_query_eval_ns_min    |  777429622100 |
| histogram_timer_rego_query_eval_ns_stddev |             0 |
+-------------------------------------------+---------------+
```